+++ src/Integration.php
@@ -10,11 +10,11 @@
 
 namespace Pronamic\WordPress\Pay\Gateways\Mollie;
 
+use Pronamic\WordPress\DateTime\DateTime;
 use Pronamic\WordPress\Pay\Core\PaymentMethods;
 use Pronamic\WordPress\Pay\AbstractGatewayIntegration;
 use Pronamic\WordPress\Pay\Payments\Payment;
 use Pronamic\WordPress\Pay\Subscriptions\Subscription as CoreSubscription;
-use WP_User;
 
 /**
  * Title: Mollie integration
@@ -145,7 +145,7 @@ class Integration extends AbstractGatewayIntegration {
 	/**
 	 * Get settings fields.
 	 *
-	 * @return array<int, array<string, array<int, string>|int|string|true>>
+	 * @return array<int, array<string, callable|int|string|bool|array<int|string,int|string>>>
 	 */
 	public function get_settings_fields() {
 		$fields = array();
@@ -272,11 +272,11 @@ class Integration extends AbstractGatewayIntegration {
 	/**
 	 * Next payment delivery date.
 	 *
-	 * @param \DateTime        $next_payment_delivery_date Next payment delivery date.
+	 * @param DateTime         $next_payment_delivery_date Next payment delivery date.
 	 * @param CoreSubscription $subscription               Subscription.
-	 * @return \DateTime
+	 * @return DateTime
 	 */
-	public function next_payment_delivery_date( \DateTime $next_payment_delivery_date, CoreSubscription $subscription ) {
+	public function next_payment_delivery_date( DateTime $next_payment_delivery_date, CoreSubscription $subscription ) {
 		$config_id = $subscription->get_config_id();
 
 		if ( null === $config_id ) {
@@ -291,7 +291,7 @@ class Integration extends AbstractGatewayIntegration {
 		}
 
 		// Check direct debit payment method.
-		$method = $subscription->get_method();
+		$method = $subscription->payment_method;
 
 		if ( null === $method ) {
 			return $next_payment_delivery_date;